requirements.txt
azure-functions
pyyaml
re
smtplib
